refactor ssr middleware
